+++ autogoal/kb/__init__.py
@@ -1,5 +1,3 @@
 from autogoal.experimental.automl import AutoML
 from autogoal.experimental.semantics import *
-from autogoal.experimental.pipeline import build_pipeline_graph, Pipeline
-
-from autogoal.kb._data import algorithm
+from autogoal.experimental.pipeline import build_pipeline_graph, Pipeline, algorithm
